add screenshots
i don't know why i chose to do this
they are all in the same workspace. why.
i'd make a really good photographer
you can see that the time is basically the same in all of these screenshots. and they're all in one workspace. that's how you know it's staged.